fix lint, create shared utils in rpc adapter
